Fix shellcheck

+++ tests/queries/0_stateless/02149_schema_inference_create_table_syntax.sh
@@ -9,7 +9,6 @@ CURDIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 USER_FILES_PATH=$(clickhouse-client --query "select _path,_file from file('nonexist.txt', 'CSV', 'val1 char')" 2>&1 | grep Exception | awk '{gsub("/nonexist.txt","",$9); print $9}')
 mkdir $USER_FILES_PATH/test_02149
 FILE_NAME=test_02149/data.Parquet
-DATA_FILE=$USER_FILES_PATH/$FILE_NAME
 
 $CLICKHOUSE_CLIENT -q "insert into table function file('$FILE_NAME', 'Parquet', 'num UInt64, str String, arr Array(UInt64)') select number as num, concat('Str: ', toString(number)) as str, [number, number + 1] as arr from numbers(10)"
 
@@ -37,5 +36,5 @@ $CLICKHOUSE_CLIENT -q "create table test_buffer engine=Buffer(currentDatabase(),
 $CLICKHOUSE_CLIENT -q "select * from test_buffer"
 $CLICKHOUSE_CLIENT -q "drop table test_buffer"
 
-rm -rf  $USER_FILES_PATH/test_02149
+rm -rf ${USER_FILES_PATH:?}/test_02149
